+++ eelbrain/pipeline.py
@@ -2,6 +2,6 @@ from ._experiment.mne_experiment import MneExperiment
 from ._experiment.preprocessing import RawSource, RawFilter, RawICA, RawMaxwell, RawReReference, RawApplyICA
 from ._experiment.epochs import PrimaryEpoch, SecondaryEpoch, SuperEpoch
 from ._experiment.groups import Group, SubGroup
-from ._experiment.parc import CombinationParc, FreeSurferParc, FSAverageParc, SeededParc, IndividualSeededParc
+from ._experiment.parc import SubParc, CombinationParc, FreeSurferParc, FSAverageParc, SeededParc, IndividualSeededParc
 from ._experiment.test_def import ANOVA, TTestOneSample, TTestInd, TTestRel, TContrastRel, TwoStageTest
 from ._experiment.variable_def import EvalVar, GroupVar, LabelVar